Question-classifies-Agno-Agent-main/Question-classifies-Agno-Agent-main/requirements.txt
agno
duckduckgo-search
tavily-python
easyocr 
asyncio
groq
python-dotenv
pandas
streamlit